Rename GenericTable file
