+++ frontend/src/common/GenericTable.tsx
